Move helpers inside data providers
