+++ x-pack/solutions/observability/plugins/apm/server/agent_builder/data_provider/get_apm_errors/index.ts
@@ -6,8 +6,8 @@
  */
 
 import datemath from '@elastic/datemath';
-import type { APMEventClient } from '../../../../lib/helpers/create_es_client/create_apm_event_client';
-import { getErrorGroupMainStatistics } from '../../../../routes/errors/get_error_groups/get_error_group_main_statistics';
+import type { APMEventClient } from '../../../lib/helpers/create_es_client/create_apm_event_client';
+import { getErrorGroupMainStatistics } from '../../../routes/errors/get_error_groups/get_error_group_main_statistics';
 import { getDownstreamServiceResource } from './get_downstream_service_resource';
 
 export async function getApmErrors(params: {
