+++ x-pack/solutions/observability/plugins/apm/server/agent_builder/data_provider/get_change_points/fetch_timeseries.ts
@@ -11,9 +11,9 @@ import type {
 } from '@elastic/elasticsearch/lib/api/types';
 import type { AggregationResultOf, AggregationResultOfMap } from '@kbn/es-types';
 import type { Unionize } from 'utility-types';
-import type { ApmDocumentType } from '../../../../../common/document_type';
-import type { RollupInterval } from '../../../../../common/rollup';
-import type { APMEventClient } from '../../../../lib/helpers/create_es_client/create_apm_event_client';
+import type { ApmDocumentType } from '../../../../common/document_type';
+import type { RollupInterval } from '../../../../common/rollup';
+import type { APMEventClient } from '../../../lib/helpers/create_es_client/create_apm_event_client';
 
 type ChangePointResult = AggregationResultOf<{ change_point: any }, unknown>;
