+++ x-pack/solutions/observability/plugins/apm/server/agent_builder/data_provider/get_change_points/get_apm_timeseries.ts
@@ -9,12 +9,12 @@ import datemath from '@elastic/datemath';
 import { kqlQuery, rangeQuery } from '@kbn/observability-plugin/server';
 import * as t from 'io-ts';
 import type { ChangePointType } from '@kbn/es-types/src';
-import { SERVICE_NAME } from '../../../../../common/es_fields/apm';
-import { LatencyAggregationType } from '../../../../../common/latency_aggregation_types';
-import { environmentQuery } from '../../../../../common/utils/environment_query';
-import { getBucketSize } from '../../../../../common/utils/get_bucket_size';
-import { termQuery } from '../../../../../common/utils/term_query';
-import type { APMEventClient } from '../../../../lib/helpers/create_es_client/create_apm_event_client';
+import { SERVICE_NAME } from '../../../../common/es_fields/apm';
+import { LatencyAggregationType } from '../../../../common/latency_aggregation_types';
+import { environmentQuery } from '../../../../common/utils/environment_query';
+import { getBucketSize } from '../../../../common/utils/get_bucket_size';
+import { termQuery } from '../../../../common/utils/term_query';
+import type { APMEventClient } from '../../../lib/helpers/create_es_client/create_apm_event_client';
 import { getErrorEventRate } from './get_error_event_rate';
 import { getExitSpanFailureRate } from './get_exit_span_failure_rate';
 import { getExitSpanLatency } from './get_exit_span_latency';
