+++ x-pack/solutions/observability/plugins/apm/server/agent_builder/data_provider/get_change_points/get_error_event_rate.ts
@@ -6,22 +6,19 @@
  */
 
 import type { QueryDslQueryContainer } from '@elastic/elasticsearch/lib/api/types';
-import { termQuery } from '@kbn/observability-plugin/server';
-import { ApmDocumentType } from '../../../../../common/document_type';
-import { TRANSACTION_NAME, TRANSACTION_TYPE } from '../../../../../common/es_fields/apm';
-import { RollupInterval } from '../../../../../common/rollup';
-import type { APMEventClient } from '../../../../lib/helpers/create_es_client/create_apm_event_client';
+import { ERROR_GROUP_NAME } from '../../../../common/es_fields/apm';
+import { ApmDocumentType } from '../../../../common/document_type';
+import { RollupInterval } from '../../../../common/rollup';
+import type { APMEventClient } from '../../../lib/helpers/create_es_client/create_apm_event_client';
 import { fetchSeries } from './fetch_timeseries';
 
-export async function getTransactionThroughput({
+export async function getErrorEventRate({
   apmEventClient,
   start,
   end,
   intervalString,
   bucketSize,
   filter,
-  transactionType,
-  transactionName,
 }: {
   apmEventClient: APMEventClient;
   start: number;
@@ -29,8 +26,6 @@ export async function getTransactionThroughput({
   intervalString: string;
   bucketSize: number;
   filter: QueryDslQueryContainer[];
-  transactionType?: string;
-  transactionName?: string;
 }) {
   const bucketSizeInMinutes = bucketSize / 60;
   const rangeInMinutes = (end - start) / 1000 / 60;
@@ -40,18 +35,20 @@ export async function getTransactionThroughput({
       apmEventClient,
       start,
       end,
-      operationName: 'assistant_get_transaction_throughput',
+      operationName: 'assistant_get_error_event_rate',
       unit: 'rpm',
-      documentType: ApmDocumentType.TransactionMetric,
-      rollupInterval: RollupInterval.OneMinute,
+      documentType: ApmDocumentType.ErrorEvent,
+      rollupInterval: RollupInterval.None,
       intervalString,
-      filter: [
-        ...filter,
-        ...termQuery(TRANSACTION_TYPE, transactionType),
-        ...termQuery(TRANSACTION_NAME, transactionName),
-      ],
-      groupByFields: [TRANSACTION_TYPE, TRANSACTION_NAME],
+      filter,
+      groupByFields: [ERROR_GROUP_NAME],
       aggs: {
+        sample: {
+          top_metrics: {
+            metrics: { field: ERROR_GROUP_NAME },
+            sort: '_score',
+          },
+        },
         value: {
           bucket_script: {
             buckets_path: {
@@ -71,10 +68,7 @@ export async function getTransactionThroughput({
   ).map((fetchedSerie) => {
     return {
       ...fetchedSerie,
-      value:
-        fetchedSerie.value !== null
-          ? (fetchedSerie.value * bucketSizeInMinutes) / rangeInMinutes
-          : null,
+      value: fetchedSerie.value !== null ? fetchedSerie.value / rangeInMinutes : null,
       data: fetchedSerie.data.map((bucket) => {
         return {
           x: bucket.key,
