+++ x-pack/solutions/observability/plugins/apm/server/agent_builder/data_provider/register_data_providers.ts
@@ -9,13 +9,13 @@ import type { CoreSetup, Logger } from '@kbn/core/server';
 import { getRollupIntervalForTimeRange } from '@kbn/apm-data-access-plugin/server/utils';
 import { getErrorSampleDetails } from '../../routes/errors/get_error_groups/get_error_sample_details';
 import { parseDatemath } from '../utils/time';
-import { getApmServiceSummary } from './helpers/get_apm_service_summary';
-import { getApmDownstreamDependencies } from './helpers/get_apm_downstream_dependencies';
+import { getApmServiceSummary } from './get_apm_service_summary';
+import { getApmDownstreamDependencies } from './get_apm_downstream_dependencies';
 import { getServicesItems } from '../../routes/services/get_services/get_services_items';
-import { getApmErrors } from './helpers/get_apm_errors';
+import { getApmErrors } from './get_apm_errors';
 import { ApmDocumentType } from '../../../common/document_type';
 import { ENVIRONMENT_ALL } from '../../../common/environment_filter_values';
-import { getExitSpanChangePoints, getServiceChangePoints } from './helpers/get_change_points';
+import { getExitSpanChangePoints, getServiceChangePoints } from './get_change_points';
 import { buildApmToolResources } from '../utils/build_apm_tool_resources';
 import type { APMPluginSetupDependencies, APMPluginStartDependencies } from '../../types';
